+++ tests/test_exec.py
@@ -1,6 +1,8 @@
 # Copyright 2025 John Sirois.
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
+from __future__ import annotations
+
 import subprocess
 import sys
 from pathlib import Path, PurePath
@@ -9,6 +11,16 @@ from textwrap import dedent
 import colors
 import pytest
 from pytest import MonkeyPatch
+
+
+@pytest.fixture(scope="session", autouse=True)
+def dev_cmd_workspace_cache_dir(tmp_path_factory) -> Path:
+    return tmp_path_factory.getbasetemp().parent / ".dev-cmd"
+
+
+@pytest.fixture(autouse=True)
+def set_dev_cmd_workspace_cache_dir(monkeypatch: MonkeyPatch, dev_cmd_workspace_cache_dir: Path):
+    monkeypatch.setenv("DEV_CMD_WORKSPACE_CACHE_DIR", dev_cmd_workspace_cache_dir.as_posix())
 
 
 @pytest.fixture
@@ -44,8 +56,26 @@ def pyproject_toml(monkeypatch: MonkeyPatch, tmp_path: Path, project_dir: PurePa
             [dependency-groups]
             dev = [
                 "ansicolors",
-                "dev-cmd @ {project_dir.as_posix()}",
+                "dev-cmd[old-pythons] @ {project_dir.as_posix()}",
+                "ruff",
             ]
+
+            [tool.dev-cmd.python.requirements]
+            # Let all the test cases share the same cache keys since requirements are hard coded
+            # here and the tests only append commands and tasks.
+
+            input-files = []
+            export-command = [
+                "uv",
+                "export",
+                "-q",
+                "--no-emit-project",
+                "--no-emit-package",
+                "dev-cmd",
+                "-o",
+                "{{requirements.txt}}",
+            ]
+            extra-requirements = ["dev-cmd[old-pythons] @ {project_dir.as_posix()}"]
             """
         )
     )
@@ -75,7 +105,17 @@ def script(tmp_path: Path) -> PurePath:
     return script
 
 
-def test_exec_python(script: PurePath, pyproject_toml: Path) -> None:
+python_args = pytest.mark.parametrize(
+    "python_args",
+    [
+        pytest.param([], id="ambient"),
+        pytest.param(["--python", sys.executable], id="custom"),
+    ],
+)
+
+
+@python_args
+def test_exec_python(script: PurePath, pyproject_toml: Path, python_args: list[str]) -> None:
     with pyproject_toml.open("a") as fp:
         fp.write(
             dedent(
@@ -90,7 +130,7 @@ def test_exec_python(script: PurePath, pyproject_toml: Path) -> None:
     assert (
         colors.cyan("Slartibartfast 42")
         == subprocess.run(
-            args=["uv", "run", "dev-cmd", "test", "--", "Slartibartfast", "42"],
+            args=["uv", "run", "dev-cmd"] + python_args + ["test", "--", "Slartibartfast", "42"],
             stdout=subprocess.PIPE,
             text=True,
             check=True,
@@ -98,7 +138,8 @@ def test_exec_python(script: PurePath, pyproject_toml: Path) -> None:
     )
 
 
-def test_exec_script(script: PurePath, pyproject_toml: Path) -> None:
+@python_args
+def test_exec_script(script: PurePath, pyproject_toml: Path, python_args: list[str]) -> None:
     with pyproject_toml.open("a") as fp:
         fp.write(
             dedent(
@@ -113,10 +154,9 @@ def test_exec_script(script: PurePath, pyproject_toml: Path) -> None:
     assert (
         colors.magenta("Ford Marvin -- Zaphod")
         == subprocess.run(
-            args=[
-                "uv",
-                "run",
-                "dev-cmd",
+            args=["uv", "run", "dev-cmd"]
+            + python_args
+            + [
                 "test",
                 "--",
                 "magenta:",
@@ -130,3 +170,17 @@ def test_exec_script(script: PurePath, pyproject_toml: Path) -> None:
             check=True,
         ).stdout.strip()
     )
+
+
+@python_args
+def test_exec_binary(script: PurePath, pyproject_toml: Path, python_args: list[str]) -> None:
+    with pyproject_toml.open("a") as fp:
+        fp.write(
+            dedent(
+                f"""
+                [tool.dev-cmd.commands]
+                check-fmt = ["ruff", "format", "--check", "{script.as_posix()}"]
+                """
+            )
+        )
+    subprocess.run(args=["uv", "run", "dev-cmd"] + python_args + ["check-fmt"], check=True)